feat(events): provide the ability to register global hammer.js events
