--- modules/@angular/platform-browser/src/dom/events/hammer_common.ts
@@ -1,55 +1,0 @@
-/**
- * @license
- * Copyright Google Inc. All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-import {EventManagerPlugin} from './event_manager';
-
-const EVENT_NAMES = {
-  // pan
-  'pan': true,
-  'panstart': true,
-  'panmove': true,
-  'panend': true,
-  'pancancel': true,
-  'panleft': true,
-  'panright': true,
-  'panup': true,
-  'pandown': true,
-  // pinch
-  'pinch': true,
-  'pinchstart': true,
-  'pinchmove': true,
-  'pinchend': true,
-  'pinchcancel': true,
-  'pinchin': true,
-  'pinchout': true,
-  // press
-  'press': true,
-  'pressup': true,
-  // rotate
-  'rotate': true,
-  'rotatestart': true,
-  'rotatemove': true,
-  'rotateend': true,
-  'rotatecancel': true,
-  // swipe
-  'swipe': true,
-  'swipeleft': true,
-  'swiperight': true,
-  'swipeup': true,
-  'swipedown': true,
-  // tap
-  'tap': true,
-};
-
-export abstract class HammerGesturesPluginCommon extends EventManagerPlugin {
-  constructor() { super(); }
-
-  supports(eventName: string): boolean {
-    return EVENT_NAMES.hasOwnProperty(eventName.toLowerCase());
-  }
-}
