+++ modules/@angular/platform-browser/src/dom/events/hammer_gestures.ts
@@ -7,7 +7,45 @@
  */
 
 import {Inject, Injectable, OpaqueToken} from '@angular/core';
-import {HammerGesturesPluginCommon} from './hammer_common';
+import {EventManagerPlugin} from './event_manager';
+
+const EVENT_NAMES = {
+  // pan
+  'pan': true,
+  'panstart': true,
+  'panmove': true,
+  'panend': true,
+  'pancancel': true,
+  'panleft': true,
+  'panright': true,
+  'panup': true,
+  'pandown': true,
+  // pinch
+  'pinch': true,
+  'pinchstart': true,
+  'pinchmove': true,
+  'pinchend': true,
+  'pinchcancel': true,
+  'pinchin': true,
+  'pinchout': true,
+  // press
+  'press': true,
+  'pressup': true,
+  // rotate
+  'rotate': true,
+  'rotatestart': true,
+  'rotatemove': true,
+  'rotateend': true,
+  'rotatecancel': true,
+  // swipe
+  'swipe': true,
+  'swipeleft': true,
+  'swiperight': true,
+  'swipeup': true,
+  'swipedown': true,
+  // tap
+  'tap': true,
+};
 
 /**
  * A DI token that you can use to provide{@link HammerGestureConfig} to Angular. Use it to configure
@@ -46,11 +84,13 @@ export class HammerGestureConfig {
 }
 
 @Injectable()
-export class HammerGesturesPlugin extends HammerGesturesPluginCommon {
+export class HammerGesturesPlugin extends EventManagerPlugin {
   constructor(@Inject(HAMMER_GESTURE_CONFIG) private _config: HammerGestureConfig) { super(); }
 
   supports(eventName: string): boolean {
-    if (!super.supports(eventName) && !this.isCustomEvent(eventName)) return false;
+    if (!EVENT_NAMES.hasOwnProperty(eventName.toLowerCase()) && !this.isCustomEvent(eventName)) {
+      return false;
+    }
 
     if (!(window as any).Hammer) {
       throw new Error(`Hammer.js is not loaded, can not bind ${eventName} event`);
