+++ modules/@angular/platform-browser/test/dom/events/key_events_spec.ts
@@ -10,7 +10,7 @@ import {describe, expect, it} from '@angular/core/testing/testing_internal';
 import {KeyEventsPlugin} from '@angular/platform-browser/src/dom/events/key_events';
 
 export function main() {
-  describe('KeyEvents', () => {
+  describe('KeyEventsPlugin', () => {
 
     it('should ignore unrecognized events', () => {
       expect(KeyEventsPlugin.parseEventName('keydown')).toEqual(null);
@@ -56,6 +56,15 @@ export function main() {
     it('should alias esc to escape', () => {
       expect(KeyEventsPlugin.parseEventName('keyup.control.esc'))
           .toEqual(KeyEventsPlugin.parseEventName('keyup.control.escape'));
+    });
+
+    it('should implement addGlobalEventListener', () => {
+      const plugin = new KeyEventsPlugin();
+
+      spyOn(plugin, 'addEventListener').and.callFake(() => {});
+
+      expect(() => plugin.addGlobalEventListener('window', 'keyup.control.esc', () => {}))
+          .not.toThrowError();
     });
 
   });
